Refactored helpers and updated error handling
