+++ app/src/main/java/com/example/mobile_development_project/data/models/UiMsgType.kt
@@ -13,5 +13,9 @@ enum class ErrorCause {
     COORDINATES_MISSING,
     BLANK_FIELDS,
     USER_ERROR,
-    GENERAL_FETCH_FAIL
+    GENERAL_FETCH_FAIL,
+    GENERAL_UPDATE_FAIL,
+    GENERAL_SAVE_FAIL,
+    TAG_EXISTS
+
 }
