+++ app/src/main/java/com/example/mobile_development_project/helpers/MapMarker.kt
@@ -1,4 +1,4 @@
-package com.example.mobile_development_project.ui
+package com.example.mobile_development_project.helpers
 
 import android.content.Context
 import android.graphics.Bitmap
